+++ src/models/stages.js
@@ -9,10 +9,6 @@ const stages = new ObjectManager({
       object.forEach(el => {
         el.dropsSet = [...el.normalDrop, ...el.extraDrop, ...el.specialDrop]
       });
-      object.push(
-        {"stageType":"ACTIVITY","stageId":"teststage01","zoneId":"testzone","code":"TE-ST-01","apCost":6,"normalDrop":["30031"],"specialDrop":[],"extraDrop":[],"dropsSet":["30031"]},
-        {"stageType":"ACTIVITY","stageId":"teststage02","zoneId":"testzone","code":"TEST-02","apCost":6,"normalDrop":["30031"],"specialDrop":[],"extraDrop":[],"dropsSet":["30031"]},
-      )
       return object
     }
   ],
