+++ src/router.js
@@ -5,7 +5,7 @@ import Home from './views/Home'
 import Report from './views/Report'
 
 import StatsLayout from './layouts/StatsLayout'
-// import StatsByStage from './views/Stats/Stage'
+import StatsByStage from './views/Stats/Stage'
 import StatsByItem from './views/Stats/Item'
 
 import AboutLayout from './layouts/AboutLayout'
@@ -17,7 +17,6 @@ import AboutContact from './views/About/Contact'
 import AboutDonate from './views/About/Donate'
 import AboutLinks from './views/About/Links'
 import NotFound from "@/views/NotFound";
-import NewStage from "@/views/Stats/NewStage";
 
 Vue.use(Router);
 
@@ -72,7 +71,7 @@ const router = new Router({
     children: [{
       path: 'stage',
       name: 'StatsByStage',
-      component: NewStage,
+      component: StatsByStage,
       props: true,
       meta: {
         icon: 'mdi-cube',
@@ -82,7 +81,7 @@ const router = new Router({
     {
       path: 'stage/:zoneId/:stageId',
       name: 'StatsByStage_Selected',
-      component: NewStage,
+      component: StatsByStage,
       props: true,
       meta: {
         hide: true,
